Proof of concept for template internationalization
The most relevant thing about the commit is that I have been able to
prepare the infrastructure for the template internationalization. The
idea is based on having a property that refers to the name of the
subfolder where to look for the templates of the language in use. It
could also have made direct use of the language of the current Locale.

There are a lot of files in the commit because I messed up with some
changes to the original templates, and I had to change a lot of files so
the application is able to pass all the tests.

There is something strange with the Message class that is executed, it
does not correspond exactly to the one in the proper package. I need
to investigate this.