🔴 Solved the problem with the generation of TikZ
This bug was driving me crazy, the LaTeX document for the
McNaughton-Yamada-Thompson exercises was given a compilation
error. After a lot of time spent trying to figure out the problem,
it was just that the String.format is Locale aware, and the
resulting strings, when the Locale was Spanish, were using the comma
instead of the dot to separate the decimal part of coordinates.